+++ src/main/java/com/jnape/palatable/lambda/io/IO.java
@@ -5,9 +5,14 @@ import com.jnape.palatable.lambda.functions.Fn1;
 import com.jnape.palatable.lambda.functor.Applicative;
 import com.jnape.palatable.lambda.monad.Monad;
 
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.Executor;
 import java.util.function.Function;
+import java.util.function.Supplier;
 
 import static com.jnape.palatable.lambda.adt.Unit.UNIT;
+import static com.jnape.palatable.lambda.functions.specialized.checked.CheckedSupplier.checked;
+import static java.util.concurrent.CompletableFuture.supplyAsync;
 
 /**
  * A {@link Monad} representing some effectful computation to be performed.
@@ -17,11 +22,37 @@ import static com.jnape.palatable.lambda.adt.Unit.UNIT;
 public interface IO<A> extends Monad<A, IO<?>> {
 
     /**
-     * Run the effect represented by this {@link IO} instance
+     * Run the effect represented by this {@link IO} instance, blocking the current thread until the effect terminates.
      *
      * @return the result of the effect
      */
     A unsafePerformIO();
+
+    /**
+     * Returns a {@link CompletableFuture} representing the result of this eventual effect. By default, this will
+     * immediately run the effect in terms of the implicit {@link Executor} available to {@link CompletableFuture}
+     * (usually the {@link java.util.concurrent.ForkJoinPool}). Note that specific {@link IO} constructions may allow
+     * this method to delegate to externally-managed {@link CompletableFuture} instead of synthesizing their own.
+     *
+     * @return the {@link CompletableFuture} representing this {@link IO}'s eventual result
+     * @see IO#unsafePerformAsyncIO(Executor)
+     */
+    default CompletableFuture<A> unsafePerformAsyncIO() {
+        return supplyAsync(this::unsafePerformIO);
+    }
+
+    /**
+     * Returns a {@link CompletableFuture} representing the result of this eventual effect. By default, this will
+     * immediately run the effect in terms of the provided {@link Executor}. Note that specific {@link IO}
+     * constructions may allow this method to delegate to externally-managed {@link CompletableFuture} instead of
+     * synthesizing their own.
+     *
+     * @return the {@link CompletableFuture} representing this {@link IO}'s eventual result
+     * @see IO#unsafePerformAsyncIO()
+     */
+    default CompletableFuture<A> unsafePerformAsyncIO(Executor executor) {
+        return supplyAsync(this::unsafePerformIO, executor);
+    }
 
     /**
      * {@inheritDoc}
@@ -115,5 +146,36 @@ public interface IO<A> extends Monad<A, IO<?>> {
      */
     static <A> IO<A> io(Fn1<Unit, A> fn1) {
         return io(() -> fn1.apply(UNIT));
+    }
+
+    /**
+     * Static factory method for creating an {@link IO} from an externally managed source of
+     * {@link CompletableFuture completable futures}.
+     * <p>
+     * Note that constructing an {@link IO} this way results in no intermediate futures being constructed by either
+     * {@link IO#unsafePerformAsyncIO()} or {@link IO#unsafePerformAsyncIO(Executor)}, and {@link IO#unsafePerformIO()}
+     * is synonymous with invoking {@link CompletableFuture#get()} on the externally managed future.
+     *
+     * @param supplier the source of externally managed {@link CompletableFuture completable futures}
+     * @param <A>      the result type
+     * @return the {@link IO}
+     */
+    static <A> IO<A> externallyManaged(Supplier<CompletableFuture<A>> supplier) {
+        return new IO<A>() {
+            @Override
+            public A unsafePerformIO() {
+                return checked(() -> unsafePerformAsyncIO().get()).get();
+            }
+
+            @Override
+            public CompletableFuture<A> unsafePerformAsyncIO() {
+                return supplier.get();
+            }
+
+            @Override
+            public CompletableFuture<A> unsafePerformAsyncIO(Executor executor) {
+                return supplier.get();
+            }
+        };
     }
 }