+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - `MapLens#asCopy` has overload taking copy function
 - `MapLens#valueAt` has overload taking copy function
 - `SortWith` for sorting an `Iterable` given a `Comparator` over its elements
+- `IO#externallyManaged`, for supplying an `IO` with externally-managed futures
 
 ### Fixed
 - issue where certain ways to compose `Effect`s unintentionally nullified the effect